python-crawler/pyproject.toml
[project]
name = "python-crawler"
version = "0.1.0"
description = "CVPR crawler for Supabase ingestion"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "beautifulsoup4>=4.12.3",
  "lxml>=5.2.1",
  "pydantic>=2.6.4",
  "python-dotenv>=1.0.1",
  "supabase>=2.4.0",
  "tenacity>=8.2.3",
  "loguru>=0.7.2",
  "certifi>=2024.2.2",
  "sentence-transformers>=3.0.1",
  "torch>=2.3.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.5.0",
]

[build-system]
requires = ["setuptools>=69.0.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.setuptools]
py-modules = [
  "config",
  "crawler",
  "logging_config",
  "models",
  "supabase_client",
  "main",
  "embedding",
]
